+++ src/wordalign.cpp
@@ -30,7 +30,7 @@ int main(int argc, char **argv)
   Log("Initialized Model1");
 
   // run Model 1 iterations
-  for (int i = 1; i <= opts.GetIterations(); i++) {
+  for (size_t i = 1; i <= opts.GetIterations(); i++) {
     model1.RunIteration(i >= opts.GetAggregateFrom());
     Log("Model1: Finished iteration " + boost::lexical_cast<string>(i));
   }
@@ -41,7 +41,7 @@ int main(int argc, char **argv)
   Log("Initialized HMM");
 
   // run HMM model iterations
-  for (int i = 1; i <= opts.GetIterations(); i++) {
+  for (size_t i = 1; i <= opts.GetIterations(); i++) {
     hmmModel.RunIteration(i >= opts.GetAggregateFrom());
     Log("HMM: Finished iteration " + boost::lexical_cast<string>(i));
   }
